refactor(common/lmlayer): drop the -word-breaker suffix
