+++ common/lexical-model-word-breakers/src/index.ts
@@ -1,4 +1,4 @@
-import {ascii} from "./ascii-word-breaker";
+import {ascii} from "./ascii";
 import {placeholder} from './placeholder';
 import {default_} from "./default";
